requirement.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
Appium-Python-Client==2.0.0
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.3.2
cryptography==41.0.5
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
pycparser==2.21
pyOpenSSL==23.3.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
requests==2.31.0
requests-mock==1.11.0
selenium==4.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
urllib3==1.26.18
urllib3-secure-extra==0.1.0
wsproto==1.2.0
